% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animals_transformed.R
\name{animals_transformed}
\alias{animals_transformed}
\title{simulate group switching to get animals_transformed df}
\usage{
animals_transformed(
  n_animals,
  n_groups,
  time_to_leave,
  time_to_return,
  travel_time,
  sampling_duration
)
}
\arguments{
\item{n_groups}{The number of modules in the network.}

\item{time_to_leave}{The average amount of time spent within the home
group prior to leaving (days per sampling period).}

\item{time_to_return}{The average amount of time spent abroad before
returning to home group.}

\item{travel_time}{Vector of the range of values that travel to the next group can take,
drawn from a uniform distribution. Default == c(0,1)}

\item{sampling_duration}{Default == 365.}
}
\description{
simulate group switching to get animals_transformed df
}
