% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bolts.R
\name{bolts}
\alias{bolts}
\title{Internal function to graph_crossing() to find new contacts}
\usage{
bolts(animal, infectious)
}
\arguments{
\item{animal}{unique individual}

\item{infectious}{a data.table with infectious individuals from previous time step}
}
\description{
Internal function to graph_crossing() to find new contacts
}
\keyword{internal}
