% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser.R
\name{browser}
\alias{browser}
\title{Environment, Module, and Pipe Browser.}
\usage{
browser(...)
}
\arguments{
\item{...}{Further arguments to be passed to \code{\link[base]{browser}}.}
}
\description{
Interrupt the execution of an expression or a pipe and allow the inspection
of the environment where browser was called from.
}
\details{
See \code{\link[base]{browser}}.
}
\examples{
\dontrun{
library(magrittr)
(function() {
  "foobar" \%>\%
    browser() \%>\%
    print
})()}

reset()
define("foo", NULL, function() {foo <- "bar"; browser()})
\dontrun{make()}

reset()
define("foobar", NULL, function() {
  library(magrittr)
  "foobar" \%>\%
    browser \%>\%
    print
})
\dontrun{make()}
}
