% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bundle.R
\name{bundle}
\alias{bundle}
\title{Bundle Module}
\usage{
bundle(...)
}
\arguments{
\item{...}{Module name as first argument. Further arguments can be passed
  for evaluation to the resulting function, if any (see \code{\link{make}}).}
}
\description{
Bundle a module and its dependencies.
}
\details{
TODO documentation
}

