% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterRank.R
\name{characterRank}
\alias{characterRank}
\title{Take a vector of characters that are numbers and return a vector of what their numeric order should be with tie-breaking}
\usage{
characterRank(x)
}
\arguments{
\item{x}{a vector of numeric characters}
}
\description{
Take a vector of characters that are numbers and return a vector of what their numeric order should be with tie-breaking
}
\keyword{internal}
