% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ids.R
\name{check_ids}
\alias{check_ids}
\title{Check that all ids are accounted for at each time step using output from simulate_schedule}
\usage{
check_ids(t2)
}
\description{
Check that all ids are accounted for at each time step using output from simulate_schedule
}
