% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{chord_engine}
\alias{chord_engine}
\title{Plot Dependencies with Chord Engine.}
\usage{
chord_engine(deps, ...)
}
\arguments{
\item{deps}{A data frame of modules and their dependencies.}

\item{...}{Further arguments to be passed to
\code{chorddiag::chorddiag}.}
}
\description{
Plot the directed acyclic graph (DAG) of modules and dependencies with a
bipartite Chord diagram.
}
\seealso{
\code{\link{plot_dependencies}}.
}
