% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_holding.R
\name{clean_holding}
\alias{clean_holding}
\title{Internal to simulate_non_independence2()}
\usage{
clean_holding(cell)
}
\arguments{
\item{cell}{the holding value string from a single cell}
}
\description{
Removes excess empty strings and dummy animals
}
\keyword{internal}
