% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{config}
\alias{config}
\alias{get}
\alias{get_all}
\alias{unset}
\alias{set}
\alias{root_config}
\alias{paths_config}
\alias{maps_config}
\title{Configurations.}
\format{
A list of getters and setters for a configuration.
 \preformatted{
 unset()
 set(..., drop = TRUE)
 get(key)
 get_all()
 }

An object of class \code{list} of length 4.

An object of class \code{list} of length 4.

An object of class \code{list} of length 4.
}
\usage{
root_config

paths_config

maps_config
}
\description{
Getters and setters for various injector configurations.
}
\section{Arguments}{

\describe{
\item{\code{...}}{A (named) list of options.}
\item{\code{key}}{A string (character vector of length one).}
\item{\code{drop}}{A flag. Should previous options be dropped and replaced?}
}
}

\examples{
reset()
root_config$get_all()
root_config$unset()
root_config$get_all()
root_config$set(c("my_modules"))
root_config$get_all()
root_config$set(c("my_great_modules"), drop = FALSE)
root_config$get_all()
}
\seealso{
\code{\link{define}}, \code{\link{get_configs}}, and \code{\link{reset}}.
}
\keyword{datasets}
