% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modulr.R
\name{define}
\alias{define}
\title{Define a module.}
\usage{
define(..., id, force = F)
}
\arguments{
\item{...}{dependencies and factory function}

\item{id}{module id}
}
\value{
module factory singleton (invisible)
}
\description{
Define a module.
}
\examples{
define(id = "module_1", function() {
 message("Module 1"); "value 1"})
define(id = "module_2", "module_1", function(m1) {
 message("Module 2 with one dependency"); paste(m1, "value 2")})
}

