% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_grp.R
\name{delta_grp}
\alias{delta_grp}
\title{Delta grp}
\usage{
delta_grp(df, column_name, value, i)
}
\arguments{
\item{df}{the t2 dataframe}

\item{column_name}{the column to reference, usually 'vector'}

\item{value}{usually the group number}

\item{i}{the row index}
}
\description{
Delta grp
}
