% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modulr.R
\name{disable_auto_redefine_and_reinstanciate}
\alias{disable_auto_redefine_and_reinstanciate}
\title{Disable module auto redefinition.}
\usage{
disable_auto_redefine_and_reinstanciate(name)
}
\description{
Disable module auto redefinition.
}

