% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modulr.R
\name{disable_auto_reinstanciate}
\alias{disable_auto_reinstanciate}
\title{Disable module auto reinstanciation.}
\usage{
disable_auto_reinstanciate(name)
}
\description{
Disable module auto reinstanciation.
}

