% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_fxn.R
\name{dt_fxn}
\alias{dt_fxn}
\title{Internal to Continuous-Time Movement Simulators}
\usage{
dt_fxn(animal)
}
\arguments{
\item{animal}{an element of animal_list from simulate_graph}
}
\description{
Takes individual time/movement descriptions and formats as a data.table
with discrete intervals
}
\keyword{internal}
