% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen_contrast.R
\name{eigen_contrast}
\alias{eigen_contrast}
\title{Eigen Contrasts}
\usage{
eigen_contrast(object, contr_object)
}
\arguments{
\item{object}{list object with trait module information}

\item{contr_object}{data frame of class `conditionContrasts`}
}
\value{
data frame
}
\description{
Eigen Contrasts
}
