% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen_cor.R
\name{eigen_cor}
\alias{eigen_cor}
\alias{ggplot_eigen_cor}
\alias{autoplot.eigen_cor}
\title{Correlation of eigentraits across responses}
\usage{
eigen_cor(object)

ggplot_eigen_cor(
  object,
  x,
  y,
  main = paste("facet by", y, "with color", x),
  ...
)

\method{autoplot}{eigen_cor}(object, ...)
}
\arguments{
\item{object}{object of class `eigen_cor`}

\item{x}{facet name}

\item{y}{color name}

\item{main}{title}

\item{...}{additional parameters}
}
\value{
data frame of class `eigen_cor`

ggplot object
}
\description{
Correlation of eigentraits across responses

GGplot of Eigentrait Correlations
}
