% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modulr.R
\name{enable_auto_redefine_and_reinstanciate}
\alias{enable_auto_redefine_and_reinstanciate}
\title{Enable module auto redefinition.}
\usage{
enable_auto_redefine_and_reinstanciate(name)
}
\description{
Enable module auto redefinition.
}

