% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modulr.R
\name{enable_auto_reinstanciate}
\alias{enable_auto_reinstanciate}
\title{Enable module auto reinstanciation.}
\usage{
enable_auto_reinstanciate(name)
}
\description{
Enable module auto reinstanciation.
}

