% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_group.R
\name{extract_group}
\alias{extract_group}
\title{Extract Group ID}
\usage{
extract_group(x)
}
\arguments{
\item{x}{a string}
}
\description{
Wrapper around stringr::str_extract with regex for group tag
}
\keyword{internal}
