% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_forward2.R
\name{ff_forward2}
\alias{ff_forward2}
\title{ff_fwd2 relates group-switching rates to average time interval lengths to let
individuals transition between home and non-home groups that they, in some cases
'magically' find (i.e., they aren't sent to home base, they're sent to their home
group as if they knew where to find them with perfect accuracy.)}
\usage{
ff_forward2(t2, curr_vec, mbrs_list, i, time_to_leave, time_to_return)
}
\description{
ff_fwd2 relates group-switching rates to average time interval lengths to let
individuals transition between home and non-home groups that they, in some cases
'magically' find (i.e., they aren't sent to home base, they're sent to their home
group as if they knew where to find them with perfect accuracy.)
}
\keyword{internal}
