% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_forward3.R
\name{ff_forward3}
\alias{ff_forward3}
\title{Fission-Fusion Forward Simulator}
\usage{
ff_forward3(t2, curr_vec, mbrs_list, i, time_to_leave, time_to_return)
}
\arguments{
\item{t2}{the data.frame of group and individual switching schedules over matched intervals}

\item{curr_vec}{the group tags present in vector at i}

\item{mbrs_list}{the list of groups and animals within groups at i}

\item{i}{the loop index for t2}

\item{time_to_leave}{The average days spent within the home
group prior to leaving.}

\item{time_to_return}{The average days spent abroad before
returning to home group.}
}
\description{
Looks forward at fission-fusion events in group-think simulations and incorporates a multinomial draw on the probability of individual switching options
}
\keyword{internal}
