% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_forward4.R
\name{ff_forward4}
\alias{ff_forward4}
\title{ff_fwd3}
\usage{
ff_forward4(t2, curr_vec, mbrs_list, i, time_to_leave, time_to_return)
}
\arguments{
\item{t2}{the dataframe of group and individual switching schedules over matched intervals}

\item{curr_vec}{the group tags present in vector at i}

\item{mbrs_list}{the list of groups and animals within groups at i}

\item{i}{the loop index for t2}

\item{time_to_leave}{The average amount of time spent within the home
group prior to leaving (days per sampling period).}

\item{time_to_return}{The average amount of time spent abroad before
returning to home group.}
}
\description{
like ff2, but trying to incorporate a multinomial draw for animals 'not-at-home'
}
\keyword{internal}
