% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_configs}
\alias{get_configs}
\title{Get Configurations.}
\usage{
get_configs()
}
\description{
Get all configurations.
}
\examples{
reset()
get_configs()

}
\seealso{
\code{\link{config}} and \code{\link{reset}}.
}
