% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/define.R
\name{get_digest}
\alias{get_digest}
\title{In order to know if a module definition has changed,
we compute a digest of it with a cryptographic hash.}
\usage{
get_digest(name)
}
\description{
In order to know if a module definition has changed,
we compute a digest of it with a cryptographic hash.
}

