% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rates.R
\name{get_rates}
\alias{get_rates}
\title{get average time to leave home and time to return home rates for every
individual in a network}
\usage{
get_rates(sched, sim)
}
\arguments{
\item{sched}{a schedule object for a network from simulate_schedule()}

\item{sim}{the simulator used to generate the schedule}
}
\description{
get average time to leave home and time to return home rates for every
individual in a network
}
\keyword{internal}
