% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/define.R
\name{get_signature}
\alias{get_signature}
\title{In order to know if a module definition has changed,
we compute a signature of it with a cryptographic hash.}
\usage{
get_signature(name)
}
\description{
In order to know if a module definition has changed,
we compute a signature of it with a cryptographic hash.
}

