% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_times.R
\name{get_times}
\alias{get_times}
\title{Calculate the Proportion of Time at Home vs. Away}
\usage{
get_times(schedule, id, simulator, option = NA)
}
\arguments{
\item{schedule}{a data.frame schedule of the type output by simulate_schedule}

\item{id}{the character string name of an individual node}

\item{simulator}{the simulator used to generate the schedule. One of: "independent", "non-independent", "group-think".}

\item{option}{additional option for group-think, "co-located" or "attached"}
}
\description{
Takes an individual's continuous-time movement description and calculates the proportion of time they spent with their home group vs any other group in their network.
}
