% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gears.R
\name{gist_gear}
\alias{gist_gear}
\title{Prepare and Publish a Modulr Gear as a Gist on Github.}
\usage{
gist_gear(name = .Last.name, load = TRUE, browse = TRUE)
}
\arguments{
\item{name}{A string (character vector of lenght one).

  A module name can contain letters, figures and some special characters,
  namely \code{_}, \code{-}, and \code{/}. The latter is a namespace
  separator.

  Names containing \code{/mock/}, \code{/mocks/}, \code{/test/},
  \code{/tests/}, \code{/example/}, or \code{/examples/} have a special
  meaning related to code testing and examples.

  The name "modulr" corresponds to a special module and is therefore
  reserved.}

\item{load}{A flag. Should an undefined module be implicitely loaded?}

\item{browse}{A flag. Should the created gist be opened in the default
browser?}
}
\description{
Prepare and publish a Modulr Gear as a Gist on Github.
}
\examples{
\dontrun{
library(gistr)
reset()
define("foo", NULL, function() NULL)
Sys.setenv("GITHUB_PAT" = "your Personal Access Token here")
gist_auth(reauth = TRUE)
gist_gear("foo")}

}
\seealso{
\code{\link{define}}, \code{\link{prepare_gear}}, and
  \code{\link{reset}}.
}

