% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make.R
\name{\%=>>\%}
\alias{\%=>>\%}
\title{Syntactic sugar to make a module and assign it in an enclosing frame.}
\usage{
lhs \%=>>\% rhs
}
\description{
Syntactic sugar to make a module and assign it in an enclosing frame.
}

