% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modulr.R
\name{\%has_default_option\%}
\alias{\%has_default_option\%}
\title{Syntactic sugar for setting default module options.}
\usage{
lhs \%has_default_option\% rhs
}
\description{
Syntactic sugar for setting default module options.
}

