% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/config.R
\name{\%has_default_options\%}
\alias{\%has_default_options\%}
\title{Syntactic sugar for setting default module options.}
\usage{
lhs \%has_default_options\% rhs
}
\description{
Syntactic sugar for setting default module options.
}

