% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modulr.R
\name{\%provides\%}
\alias{\%provides\%}
\title{Syntactic sugar to provide a factory, can be used in conjunction with \%requires\%.}
\usage{
lhs \%provides\% rhs
}
\description{
Syntactic sugar to provide a factory, can be used in conjunction with \%requires\%.
}

