% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modulr.R
\name{\%requires\%}
\alias{\%requires\%}
\title{Syntactic sugar to require dependencies, to be used in conjunction with \%provides\%.}
\usage{
lhs \%requires\% rhs
}
\description{
Syntactic sugar to require dependencies, to be used in conjunction with \%provides\%.
}

