% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_from_at.R
\name{graph_from_at}
\alias{graph_from_at}
\title{Simulate the true simple ratio index values for each dyad in a network
using output from animals_transformed().}
\usage{
graph_from_at(animals_transformed, animal_list)
}
\arguments{
\item{animals_transformed}{output from the animals_transformed function}

\item{animal_list}{output from the at fxn}
}
\value{
sim_igraph, an igraph graph object.
}
\description{
Simulate the true simple ratio index values for each dyad in a network
using output from animals_transformed().
}
