% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_from_non_independence.R
\name{graph_from_non_independence}
\alias{graph_from_non_independence}
\title{graph non independence}
\usage{
graph_from_non_independence(t2)
}
\arguments{
\item{t2}{a dataframe from simulate_non_independence()}
}
\description{
graph non independence
}
