% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_module}
\alias{import_module}
\alias{\%digests\%}
\alias{\%imports\%}
\title{Import a Module.}
\usage{
import_module(name, url, ..., digest = NULL, force = FALSE)
}
\arguments{
\item{name}{A string (character vector of lenght one).

  A module name can contain letters, figures and some special characters,
  namely \code{_}, \code{-}, and \code{/}. The latter is a namespace
  separator.

  Names containing \code{/mock/}, \code{/mocks/}, \code{/test/},
  \code{/tests/}, \code{/example/}, or \code{/examples/} have a special
  meaning related to code testing and examples.

  The name "modulr" corresponds to a special module and is therefore
  reserved.}

\item{url}{A string (character vector of length one). The URL must use the
HTTP(S) protocol.}

\item{...}{Further arguments passed to \code{httr::\link[httr]{GET}}.}

\item{digest}{A string (character vector of length one). See
\code{\link{get_digest}}.}

\item{force}{A flag. Should an already defined module be re-imported?}
}
\value{
The result of the evaluation of the imported script.
}
\description{
Import or re-import a module which is defined in an R, R Markdown or R Sweave
script at a given URL using the HTTP(S) protocol.
}
\details{
R Markdown and R Sweave files are accordingly tangled into R code, which is
in turn evaluated.

 The imported module is rejected if
 \itemize{
   \item its name differs from the \code{name} argument,
   \item its digest differs from the \code{digest} argument.
 }
 In such a case, the internal state of modulr is rolled back.
}
\section{Syntactic Sugars}{

 \preformatted{name \%imports\% url}
 \preformatted{name \%digests\% digest \%imports\% url}
}

\section{Warning}{

 It is considered a very bad practice to define, touch, undefine, load, make,
 reset, or perform any other operation from within a module definition that
 may alterate the internal state of modulr.
}

\examples{
\dontrun{
reset()
# https://gist.github.com/aclemen1/3fcc508cb40ddac6c1e3
"modulr/vault" \%imports\%
  "https://gist.github.com/aclemen1/3fcc508cb40ddac6c1e3"
list_modules()
reset()
# equivalently
"modulr/vault" \%imports\% "3fcc508cb40ddac6c1e3"
make("modulr/vault/example")
make_tests()}

\dontrun{
reset()
# https://gist.github.com/aclemen1/3fcc508cb40ddac6c1e3
"modulr/vault_with_a_typo" \%imports\% "3fcc508cb40ddac6c1e3"
list_modules()}

\dontrun{
reset()
# https://gist.github.com/aclemen1/3fcc508cb40ddac6c1e3
"modulr/vault" \%digests\%
  "with a wrong digest" \%imports\%
  "3fcc508cb40ddac6c1e3"
list_modules()}

}
\seealso{
\code{httr::\link[httr]{GET}}, \code{\link{get_digest}},
 \code{\link{list_modules}}, \code{\link{make}}, and
 \code{\link{make_tests}}.
}
