% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_back.R
\name{index_back}
\alias{index_back}
\title{Internal to simulate_non_independence}
\usage{
index_back(df, column_name, value, i)
}
\arguments{
\item{df}{the t2 dataframe}

\item{column_name}{the column to reference, usually 'vector'}

\item{value}{usually the group number}

\item{i}{the row index}
}
\description{
Pulls the row number of the last occurrence of a group from any combination of vector values
}
\keyword{internal}
