% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_back_same.R
\name{index_back_same}
\alias{index_back_same}
\title{Pulls the row number of the last occurrence of a group}
\usage{
index_back_same(df, column_name, value, i)
}
\arguments{
\item{df}{the t2 dataframe}

\item{column_name}{the column to reference, usually 'vector'}

\item{value}{usually the group number}

\item{i}{the row index}
}
\description{
Pulls the row number of the last occurrence of a group
}
\keyword{internal}
