% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_forward.R
\name{index_forward}
\alias{index_forward}
\title{Index forward}
\usage{
index_forward(df, column_name, value, i)
}
\arguments{
\item{df}{the t2 dataframe}

\item{column_name}{the column to reference, usually 'vector'}

\item{value}{usually the group number}

\item{i}{the row index}
}
\description{
Index forward
}
