% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate_group.R
\name{initiate_group}
\alias{initiate_group}
\title{Extract data from list on individual basis and make keyed data.table}
\usage{
initiate_group(x, n)
}
\description{
Extract data from list on individual basis and make keyed data.table
}
