% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modulr.R
\name{instanciate}
\alias{instanciate}
\title{Instanciate a module.}
\usage{
instanciate(name, force_reinstanciate = F, force_redefine_reinstanciate = F,
  force_reinstanciate_all = F, force_redefine_reinstanciate_all = F)
}
\description{
Instanciate a module.
}

