% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_function.R
\name{interval_function}
\alias{interval_function}
\title{Extract data from list on individual basis and make keyed data.table}
\usage{
interval_function(groups, intervals)
}
\arguments{
\item{groups}{a pairwise set of groups}

\item{intervals}{the intervals dataframe}
}
\description{
Extract data from list on individual basis and make keyed data.table
}
