% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulr.R
\name{.Last.name}
\alias{.Last.name}
\title{Name of Last Used Module.}
\usage{
.Last.name
}
\value{
The name of the last used module.
}
\description{
The name of the last used module.
}
\details{
A module is considered used last if it has just been:
\itemize{
 \item defined or imported with a name without special meaning and not
   reserved (see \code{\link{define}}),
 \item made or touched with a non-reserved name.
}
Other actions on modules do not alter the value of \code{.Last.name}.
}
\section{Warning}{

Do not assign to \code{.Last.name} in the workspace, because this will always
mask the object of the same name in \code{package:modulr}.
}

\examples{
define("foo", NULL, function() NULL)
.Last.name
define("bar/test", NULL, function() NULL)
.Last.name
make("bar/test")
.Last.name
}
\seealso{
\code{\link{define}}.
}
