% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_modules}
\alias{list_modules}
\alias{lsmod}
\title{List Defined Modules.}
\usage{
list_modules(
  regexp,
  reserved = FALSE,
  wide = TRUE,
  full = FALSE,
  formatted = TRUE,
  absolute = FALSE,
  cols = c("name", "version", "storage", "along", "type", "weight", "calls",
    "dependencies", "uses", "size", "lines", "modified")
)

lsmod(
  regexp,
  reserved = FALSE,
  wide = TRUE,
  full = FALSE,
  formatted = TRUE,
  absolute = FALSE,
  cols = c("name", "version", "storage", "along", "type", "weight", "calls",
    "dependencies", "uses", "size", "lines", "modified")
)
}
\arguments{
\item{regexp}{A regular expression. If not missing, the regular expression
is used to filter the names of the modules to be made.}

\item{reserved}{A flag. Should special modules with a reserved name be
considered?}

\item{wide}{A flag. Should the function return a data frame instead of a
characted vector?}

\item{full}{A flag. Should all columns be included in the data frame?}

\item{formatted}{A flag. Should columns with units be string formatted?}

\item{absolute}{A flag. Should the file paths be absolute?}

\item{cols}{A character vector. Details returned in the columns of the data
frame.}
}
\value{
A character vector or a data frame containing module informations.
}
\description{
List defined modules.
}
\details{
For each module, the following details can be returned in the columns of the
data frame:
\describe{
\item{\code{name}}{name.}
\item{\code{version}}{version of the module.}
\item{\code{storage}}{storage of the module, `in-memory` or `on-disk`.}
\item{\code{along}}{name of the on-disk module along which the definition
takes place.}
\item{\code{filepath}}{file path of the `on-disk` module.}
\item{\code{type}}{type of the object returned.}
\item{\code{weight}}{memory size of the object.}
\item{\code{calls}}{number of explicit make calls.}
\item{\code{dependencies}}{number of direct dependencies.}
\item{\code{uses}}{number of modules requiring the module.}
\item{\code{size}}{memory size occupied by the definition.}
\item{\code{lines}}{number of lines of the provider.}
\item{\code{chars}}{number of characters of the provider.}
\item{\code{duration}}{duration of the evaluation.}
\item{\code{modified}}{timestamp of last modification.}
\item{\code{created}}{timestamp of creation.}
\item{\code{digest}}{digest (cf. \code{\link{get_digest}}).}
}
}
\examples{
reset()
list_modules()
define("foo", NULL, function() Sys.sleep(1L))
list_modules()
list_modules(reserved = TRUE)
list_modules(reserved = TRUE, wide = FALSE)
invisible(make("foo"))
list_modules(reserved = TRUE, full = TRUE)
list_modules(
  reserved = TRUE,
  formatted = FALSE,
  cols = c("weight", "size", "modified", "created"))
define("bar", NULL, function() Sys.sleep(1L))
define("foobar", list(f = "foo", b = "bar"), function(f, b) NULL)
invisible(make("foobar"))
Sys.sleep(1L)
touch("foo")
list_modules(".oo.*", cols = c("weight", "size", "modified", "created"))

}
\seealso{
\code{\link{define}}, \code{\link{make}},
  \code{\link{reset}}, and \code{\link{touch}}.
}
