% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulR.r
\name{m_arithmetic}
\alias{m_arithmetic}
\title{Function factory for doing modular arithmetic}
\usage{
m_arithmetic(z)
}
\arguments{
\item{z}{The ring size.}
}
\description{
Function factory for doing modular arithmetic
}
