% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulR.r
\name{m_table}
\alias{m_table}
\title{Create a arithmetic table in Z_n space}
\usage{
m_table(z, FUN = "*")
}
\arguments{
\item{z}{The ring size.}

\item{FUN}{The binary function to be applied.}
}
\description{
Create a arithmetic table in Z_n space
}
