% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modulr.R
\docType{data}
\name{maps_config}
\alias{maps_config}
\title{Maps configuration.}
\format{\preformatted{List of 3
 $ set    :function (..., drop = T)  
 $ get    :function (key)  
 $ get_all:function ()  
}}
\usage{
maps_config
}
\description{
Maps configuration.
}
\keyword{datasets}

