% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_factors.R
\name{module_factors}
\alias{module_factors}
\title{Extract module information from kME and colors}
\usage{
module_factors(kME, colors)
}
\arguments{
\item{kME}{data frame of kME correlations of traits with eigentraits}

\item{colors}{color for modules represented by eigentraits}
}
\value{
data frame with trait, module, kME
}
\description{
Extract module information from kME and colors
}
