% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_kMEs.R
\name{module_kMEs}
\alias{module_kMEs}
\alias{ggplot_module_kMEs}
\alias{autoplot.module_kMEs}
\title{Extract Modules and kMEs}
\usage{
module_kMEs(object)

ggplot_module_kMEs(
  object,
  x,
  y,
  abskME = FALSE,
  title = paste("facet by", y, "with color by", x),
  ...
)

\method{autoplot}{module_kMEs}(object, ...)
}
\arguments{
\item{object}{object of class `module_kMEs`}

\item{x}{name of x response}

\item{y}{name of y response}

\item{abskME}{plot absolute values if `TRUE`}

\item{title}{title of plot}

\item{...}{additional parameters}
}
\value{
object of class `module_kMEs`

ggplot object
}
\description{
Extract Modules and kMEs

GGplot of Module kMEs
}
