% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{module_options}
\alias{module_options}
\alias{\%has_default_option\%}
\alias{\%has_default_options\%}
\alias{\%has_option\%}
\alias{\%has_options\%}
\title{Module Options.}
\usage{
module_options(name = .Last.name)

name \%has_default_option\% options

name \%has_default_options\% options

name \%has_option\% options

name \%has_options\% options
}
\arguments{
\item{name}{A string (character vector of lenght one).

  A module name can contain letters, figures and some special characters,
  namely \code{_}, \code{-}, and \code{/}. The latter is a namespace
  separator.

  Names containing \code{/mock/}, \code{/mocks/}, \code{/test/},
  \code{/tests/}, \code{/example/}, or \code{/examples/} have a special
  meaning related to code testing and examples.

  The name "modulr" corresponds to a special module and is therefore
  reserved.}

\item{options}{A (named) list of options.}
}
\value{
A list of getters and setters. See \code{\link{config}}.
}
\description{
Getters and setters for module options.
\bold{Deprecated and kept for backward compatibility.}
}
\section{Syntactic Sugars}{

 \code{name \%has_default_option\% options} and
 \code{name \%has_default_options\% options} for
   \code{module_options(name)$set(options, drop = FALSE)}.

 \code{name \%has_option\% options} and
 \code{name \%has_options\% options} for
   \code{module_options(name)$set(options, drop = TRUE)}.
}

\section{Warning}{

 It is considered a very bad practice to define, touch, undefine, load, make,
 reset, or perform any other operation from within a module definition that
 may alterate the internal state of modulr.


Deprecated and kept for backward compatibility.
}

