% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modulr.R
\docType{package}
\name{modulr}
\alias{.modulr}
\alias{modulr}
\alias{modulr-package}
\title{modulr -- A Module Pattern for R}
\format{\preformatted{List of 7
 $ define    :function (..., id, force = F)  
 $ run       :function (...)  
 $ .eval     :function (id, force = F)  
 $ configure :function (config)  
 $ reset     :function ()  
 $ undef     :function (id)  
 $ .internals:function ()  
}}
\usage{
.modulr
}
\description{
modulr is an R implementation of the Module Pattern that allows you to encapsulate pieces of code into useful singleton units, namely modules that register their capabilities, export values and rely on other modules dependencies. modulr is widely inspired from RequireJS for Javascript and the work of James Burke and other contributors.
}
\author{
Alain Clément-Pavon <\email{alain.clement-pavon@unil.ch}>
}
\keyword{datasets}

