% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcarto_modules.R
\name{netcarto_modules}
\alias{netcarto_modules}
\title{Return Netcarto Membership Vector}
\usage{
netcarto_modules(x)
}
\arguments{
\item{x}{a named square adjacency matrix or named 'igraph' graph. Intended for
weighted, undirected graphs.}

\item{alg}{the community detection algorithm to use, either from igraph:
"fast_greedy", "leading_eigen", "leiden", "louvain", or "walktrap"
or from rnetcarto: "netcarto"
default is "walktrap"}
}
\value{
a named vector of module memberships matching the named adjacency
matrix or igraph graph object (igraph::V(graph)$name)
}
\description{
Convenience function to extract a membership vector using the netcarto community
detection algorithm.
}
\examples{
\donttest{
adjmat <- matrix(sample(seq(0, 1, .01), 16), nrow = 4)
diag(adjmat) <- 0
adjmat[lower.tri(adjmat)] <- 0
rownames(adjmat) <- colnames(adjmat) <- paste0("Animal_", 1:4)
netcarto_modules(adjmat)
}
}
