% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_dt_fxn.R
\name{new_dt_fxn}
\alias{new_dt_fxn}
\title{Internal to sglt_bind}
\usage{
new_dt_fxn(obj_in_sgl)
}
\arguments{
\item{obj_in_sgl}{a dataframe from sub_group_list}
}
\description{
Helper to handle sub-group splitting
}
\keyword{internal}
