% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special.R
\name{options_provider}
\alias{options_provider}
\alias{\%provides_options\%}
\title{Module Options Provider}
\usage{
options_provider(...)
}
\arguments{
\item{...}{Named arguments used as default options. If void, no default
option is set.}
}
\value{
A module provider which exposes an R environment. The default options
  arguments, if any, are used to assign the values to their corresponding
  names in the returned environment.
}
\description{
Construct a module provider intended for defining modules that can be used as
general purpose options and configuration containers.
}
\section{Syntactic Sugars}{

 \preformatted{name \%provides_options\% options}
}

\section{Warning}{

 It is considered a very bad practice to define, touch, undefine, load, make,
 reset, or perform any other operation from within a module definition that
 may alterate the internal state of modulr.
}

\examples{
reset()

"foo/config" \%provides\% options_provider(upper = FALSE)

"foo" \%requires\% list(config = "foo/config") \%provides\% {
  function() casefold("foo", upper = config$upper)
}

foo \%<=\% "foo"
foo()

config \%<=\% "foo/config"
config$upper <- TRUE
foo()

"foo/config" \%provides_options\% list(upper = FALSE)
foo \%<=\% "foo"
foo()

touch("foo/config")
"foo/config" \%provides_options\% list(upper = FALSE)
foo \%<=\% "foo"
foo()

}
\seealso{
\code{\link{\%<=\%}}, \code{\link{define}}, \code{\link{make}},
  \code{\link{\%provides\%}}, \code{\link{\%requires\%}},
  \code{\link{reset}}, and \code{\link{touch}}.
}
