% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special.R
\name{options_provider_}
\alias{options_provider_}
\title{Options module which exposes an R environment. See ?options_provider.}
\usage{
options_provider_()
}
\description{
Options module which exposes an R environment. See ?options_provider.
}
