% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modulr.R
\docType{data}
\name{paths_config}
\alias{paths_config}
\title{Paths configuration.}
\format{\preformatted{List of 3
 $ set    :function (..., drop = T)  
 $ get    :function (key)  
 $ get_all:function ()  
}}
\usage{
paths_config
}
\description{
Paths configuration.
}
\keyword{datasets}

