% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_dependencies}
\alias{graph_dependencies}
\alias{plot_dependencies}
\title{Plot Dependencies.}
\usage{
plot_dependencies(group, reserved = TRUE)
}
\arguments{
\item{group}{A character vector of module names (cf. \code{\link{define}}) to
  include as a subset of the graph nodes.}

\item{reserved}{A flag. Should special modules with a reserved name be
considered?}
}
\description{
Plot the directed acyclic graph (DAG) of modules and dependencies.
}
\examples{
reset()
define("foo", NULL, function() NULL)
define("bar", list(m = "modulr"), function(m) NULL)
define("foobar", list(f = "foo", b = "bar"), function(f, b) NULL)
define("foobuzz", list(f = "foo", b = "buzz"), function(f, b) NULL)
define("fizz", list(f1 = "foobar", f2 = "foobuz"), function(f1, f2) NULL)
wait <- function() invisible(readline(prompt="Press [enter] to continue"))
plot_dependencies("foobar", reserved = FALSE); wait()
plot_dependencies("foobar"); wait()
plot_dependencies(reserved = FALSE); wait()
plot_dependencies()
}
\seealso{
\code{\link{define}} and \code{\link{reset}}.
}

