% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_non_independence.R
\name{plot_non_independence}
\alias{plot_non_independence}
\title{plot non independence}
\usage{
plot_non_independence(
  g,
  vertex.size = 40,
  mark.expand = 25,
  vertex.label = "",
  vertex.label.cex = 1.5,
  title = ""
)
}
\arguments{
\item{g}{an igraph object output from simulate_graph()}

\item{vertex.size}{node size in plot}

\item{mark.expand}{padding of polygon denoting modules around nodes}

\item{vertex.label}{label for graph vertices}

\item{vertex.label.cex}{size for vertex labels}

\item{title}{plot title}
}
\description{
plot non independence
}
