% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_null.R
\name{plot_null}
\alias{plot_null}
\title{Plot Null Message Graphic}
\usage{
plot_null(msg = "no data")
}
\arguments{
\item{msg}{character string message to display}
}
\value{
ggplot object
}
\description{
Displays a void ggplot graphic with centered text when data is missing or empty.
}
