% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sampled_graph.R
\name{plot_sampled_graph}
\alias{plot_sampled_graph}
\title{plot_sampled_graph()}
\usage{
plot_sampled_graph(g_obs, g, vertex.size = 40, mark.expand = 25)
}
\arguments{
\item{g_obs, }{an igraph object output from sample_graph()}

\item{g}{an igraph object output from simulate_graph()}
}
\value{

}
\description{
plot_sampled_graph()
}
\examples{
set.seed(123)

g <- simulate_graph(n_animals = 25,
                    n_groups = 4,
                    time_to_leave = 5,
                    time_to_return = 2,
                    samples_per_day = 1,
                    sampling_duration = 7,
                    time_cut = 7)
igraph::V(g)$name <- stringr::str_extract(igraph::V(g)$name, "\\\\d{1,}")

g_obs <- sample_graph(
  graph = g,
  missingness = 0.5,
  propGPS = 1,
  regime = "better")

par(mfrow = c(1,2))
plot_simulated_graph(g)
plot_sampled_graph(g_obs, g)

}
\seealso{
\code{\link{sample_graph, simulate_graph, plot_simulated_graph}}
}
