% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_simulated_graph.R
\name{plot_simulated_graph}
\alias{plot_simulated_graph}
\title{Plot Graph from Continuous-Time Movement Descriptions}
\usage{
plot_simulated_graph(
  g,
  vertex.size = 40,
  mark.expand = 25,
  vertex.label = NA,
  vertex.label.cex = 1.5,
  title = "",
  seed = NULL
)
}
\arguments{
\item{g}{an igraph object output from simulate_graph() or graph_from_schedule()}

\item{vertex.size}{node size in plot}

\item{mark.expand}{padding of polygon denoting modules around nodes}

\item{vertex.label}{label for graph vertices}

\item{vertex.label.cex}{size for vertex labels}

\item{title}{plot title}

\item{seed}{optional integer value to set.seed() within function and preserve node layout when plotting sampled and simulated graphs}
}
\value{
a plot of the igraph object
}
\description{
Plot a graph from modulr graphing functions. Optional seed can be set to
facilitate side-by-side comparison with plot_sampled_graph.
}
\examples{
\donttest{
g <- simulate_graph(n_animals = 25,
                    n_groups = 4,
                    time_to_leave = 5,
                    time_to_return = 2,
                    travel_time = c(0.01, 0.2),
                    samples_per_day = 1,
                    sampling_duration = 7)
plot_simulated_graph(g)
}
}
