% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{print_info}
\alias{print_info}
\alias{docstring}
\alias{info}
\title{Output Docstring Info.}
\usage{
print_info(name = .Last.name, line_numbers = FALSE, sep = "\\n", load = TRUE)

info(name = .Last.name, line_numbers = FALSE, sep = "\\n", load = TRUE)
}
\arguments{
\item{name}{A string (character vector of lenght one).

  A module name can contain letters, figures and some special characters,
  namely \code{_}, \code{-}, and \code{/}. The latter is a namespace
  separator.

  Names containing \code{/mock/}, \code{/mocks/}, \code{/test/},
  \code{/tests/}, \code{/example/}, or \code{/examples/} have a special
  meaning related to code testing and examples.

  The name "modulr" corresponds to a special module and is therefore
  reserved.}

\item{line_numbers}{A flag. Should the source line numbers be outputed?}

\item{sep}{A string (character vector of length one) containing the separator
between docstrings blocs.}

\item{load}{A flag. Should the module be loaded?}
}
\description{
Output the docstring info of a module.
}
\details{
A docstring is intended to document a module and provide the user
with the ability to inspect it at run time, for instance as an interactive
help system, or as metadata. Formally, it is a block of commented lines
prefixed with \code{#'} and located at the top of the module provider.

The preferred formatting for a docstring is R Markdown, notabely for Modulr
Gears (see \code{\link{prepare_gear}}).
}
\examples{
define("foo", NULL, {
 #' # Info
 #' This is a docstring for `foo`.
})
print_info("foo")

\dontrun{
tmp <- tempfile(fileext = ".html")
cat((knitr::knit2html(text = info("foo"))), file = tmp)
if(interactive()) rstudio::viewer(tmp)
Sys.sleep(1L); unlink(tmp)}
}
\seealso{
\code{\link{define}}, \code{\link{prepare_gear}}, and
  \code{\link{reset}}.
}
