% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_rel.R
\name{q_rel}
\alias{q_rel}
\title{Calculate the assortativity coefficient on a weighted, undirected network graph}
\usage{
q_rel(graph_object)
}
\arguments{
\item{graph_object}{an \code{igraph} graph object with 'membership' attribute}
}
\description{
Calculate the assortativity coefficient on a weighted, undirected network graph
}
