% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_vect.R
\name{rand_vect}
\alias{rand_vect}
\title{Internal to simulate_non_independence}
\usage{
rand_vect(N, M, sd = 1, pos.only = TRUE)
}
\description{
Generates multinomial draws for group-switching
}
\keyword{internal}
