% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breadcrumbs.R
\name{reactivate_breadcrumbs}
\alias{reactivate_breadcrumbs}
\title{Re-Activate Breadcrumbs.}
\usage{
reactivate_breadcrumbs()
}
\description{
Re-activate breadcrumbs messages on error.
}
\details{
Breadcrumbs messages are activated by default when the modulr package is
loaded. This is done by binding a wrapper function to the
\code{\link[=options]{error}} option. The function wraps any other previously
binded function and the \code{\link{get_breadcrumbs}} function together.
Since IDEs like RStudio are likely to reset this option during the lifetime
of an R session, \code{reactivate_breadcrumbs} can be useful to re-install
the wrapper function.
}
\examples{
reset()
options(error = browser)
reactivate_breadcrumbs()
define("foo", NULL, function() function() stop("error in 'foo'"))
define("bar", list(foo = "foo"), function(foo) function() foo())
define("foobar", list(bar = "bar"), function(bar) bar())
\dontrun{make()}

}
