% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real_networks.R
\docType{data}
\name{real_networks}
\alias{real_networks}
\title{Simulated networks with 25 nodes and 2 to 5 modules.}
\format{
A list of 24 igraph objects.
}
\usage{
real_networks
}
\description{
A list of igraph objects. Time to leave group and time to return to group
take values (5,10) and (1,3,5,7), for all combinations where time to leave >
time to return.
}
\keyword{datasets}
