% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rejoin_animals.R
\name{rejoin_animals}
\alias{rejoin_animals}
\title{rejoin animals to their home groups}
\usage{
rejoin_animals(inlist, elem_rejoin)
}
\arguments{
\item{inlist}{a list like mbrs_list}

\item{elem_rejoin}{a list with same length  as mbrs_list of animals to rejoin
their home group within current groups}
}
\description{
rejoin animals to their home groups
}
\keyword{internal}
