% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relocate_animals.R
\name{relocate_animals}
\alias{relocate_animals}
\title{Internal to simulate_non_independence}
\usage{
relocate_animals(inlist, elem_relocate)
}
\arguments{
\item{inlist}{a list like mbrs_list}

\item{elem_relocate}{a list with same length as mbrs_list of animals to send
to home groups at next time slot}
}
\description{
Relocate animals going home to a new mbrs_list slot
}
\keyword{internal}
