% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_animals.R
\name{remove_animals}
\alias{remove_animals}
\title{Deletes animals from a nested list}
\usage{
remove_animals(inlist, elem_remove)
}
\arguments{
\item{inlist}{a list like mbrs_list}

\item{elem_remove}{a list with same length  as mbrs_list of animals to delete}
}
\description{
Deletes animals from a nested list
}
\keyword{internal}
