% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modulr.R
\name{run}
\alias{run}
\title{Run a module.}
\usage{
run(...)
}
\arguments{
\item{...}{dependencies and factory function}

\item{id}{module id}
}
\value{
module factory singleton (invisible)
}
\description{
Run a module.
}

