% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/run.R
\name{run_async}
\alias{run_async}
\title{Run a Module Asynchronously.}
\usage{
run_async(...)
}
\arguments{
\item{...}{Module name as first argument. Further arguments can be passed
  for evaluation to the resulting function, if any (see \code{\link{make}}).}
}
\description{
Run a module asynchronously as a standalone script.
}
\details{
TODO documentation
}

