% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_vec.R
\name{sample_vec}
\alias{sample_vec}
\title{Custom function to sample from integers}
\usage{
sample_vec(x, ...)
}
\arguments{
\item{x}{one or more integers to sample from}
}
\description{
Used in simulate_animal and simulate_groups function to avoid issues when
there are only two groups and animals_other_groups is a length 1 integer
}
\keyword{internal}
