% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sankey_engine}
\alias{sankey_engine}
\title{Plot Dependencies with Sankey Engine.}
\usage{
sankey_engine(deps, ...)
}
\arguments{
\item{deps}{A data frame of modules and their dependencies.}

\item{...}{Further arguments to be passed to
\code{networkD3::\link[networkD3]{sankeyNetwork}}.}
}
\description{
Plot the directed acyclic graph (DAG) of modules and dependencies with a
Sankey diagram.
}
\seealso{
\code{\link{plot_dependencies}}.
}
