% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrub_dummies.R
\name{scrub_dummies}
\alias{scrub_dummies}
\title{drop unnecessary dummy animals from a cell}
\usage{
scrub_dummies(cell)
}
\arguments{
\item{cell}{a cell containing a string of ids}
}
\description{
drop unnecessary dummy animals from a cell
}
\keyword{internal}
