% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulr.R
\name{set_verbosity}
\alias{set_verbosity}
\alias{get_verbosity}
\title{Set and Get Verbosity Level.}
\usage{
set_verbosity(level = 2L)

get_verbosity()
}
\arguments{
\item{level}{A scalar (integer vector of length one),
possibly \code{-Inf} or \code{+Inf}.}
}
\description{
Set and get the global verbosity level.
}
\details{
Messages are generated, accordingly to the following levels:

\describe{
\item{Level 0.}{None.}
\item{Level 1.}{Operations modifying the internal state.}
\item{Level 2 (default).}{All operations.}
\item{Level \code{+Inf}.}{Debugging informations.}
}
}
\examples{
reset()
set_verbosity(+Inf)
define("foo", NULL, function() "Hello World")
define("bar", list(f = "foo"), function(f) sprintf("*\%s*", f))
make()
touch("foo")
make("bar")

reset()
set_verbosity(2)
define("foo", NULL, function() "Hello World")
define("bar", list(f = "foo"), function(f) sprintf("*\%s*", f))
make()
touch("foo")
make("bar")

reset()
set_verbosity(1L)
define("foo", NULL, function() "Hello World")
define("bar", list(f = "foo"), function(f) sprintf("*\%s*", f))
make()
touch("foo")
make("bar")

reset()
set_verbosity(0L)
define("foo", NULL, function() "Hello World")
define("bar", list(f = "foo"), function(f) sprintf("*\%s*", f))
make()
touch("foo")
make("bar")

}
\seealso{
\code{\link{define}}, \code{\link{make}}, \code{\link{reset}}, and
  \code{\link{touch}}.
}
