% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulr.R
\name{.SharedEnv}
\alias{.SharedEnv}
\alias{sharedenv}
\title{Shared Environment.}
\usage{
.SharedEnv

sharedenv()
}
\value{
The shared environment.
}
\description{
A shared environment between modules.
}
\details{
The shared environment is bound to the injector being used.
}
\section{Warning}{

Do not assign to \code{.SharedEnv} in the workspace, because this will always
mask the object of the same name in \code{package:modulr}.
}

\examples{
.SharedEnv$foo <- "foo"
"foobar" \%provides\% { function() print(.SharedEnv$foo) }
make()()
.SharedEnv$foo <- "bar"
make()()
}
\seealso{
\code{\link{injector}}.
}
