% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_animal.R
\name{simulate_animal}
\alias{simulate_animal}
\title{Simulate an Individual Animal}
\usage{
simulate_animal(
  n_groups,
  time_to_leave,
  time_to_return,
  travel_time = c(0.001, 0.002),
  sampling_duration,
  samples_per_day = 1
)
}
\arguments{
\item{n_groups}{The number of modules in the network.}

\item{time_to_leave}{The average days spent within the home
group prior to leaving.}

\item{time_to_return}{The average days spent abroad before
returning to home group.}

\item{travel_time}{Vector of the range of values that travel to the next group can take,
drawn from a uniform distribution.}

\item{sampling_duration}{The number of days to simulate.}

\item{samples_per_day}{Only used in discrete time approximations in simulate_graph.}
}
\value{
A vector of sampling locations for a single animal.
}
\description{
Simulates independent movement behavior for a single animal.
}
\examples{
\donttest{
animal_one <- simulate_animal(n_groups = 4,
 time_to_leave = 5,
 time_to_return = 2,
 travel_time = c(0.001, 0.002),
 sampling_duration = 30)
}
}
