% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_graph.R
\name{simulate_graph}
\alias{simulate_graph}
\title{Simulate Independent Switching Graph}
\usage{
simulate_graph(
  n_animals,
  n_groups,
  time_to_leave,
  time_to_return,
  travel_time,
  sampling_duration,
  sampler = "discrete",
  samples_per_day = 1
)
}
\arguments{
\item{n_animals}{The number of nodes to include in the network.}

\item{n_groups}{The number of modules in the network.}

\item{time_to_leave}{The average days spent within the home
group prior to leaving.}

\item{time_to_return}{The average days spent abroad before
returning to home group.}

\item{travel_time}{Vector of the range of values that travel to the next group can take,
drawn from a uniform distribution.}

\item{sampling_duration}{The number of days to simulate.}

\item{sampler}{Takes one of c("discrete", "continuous"). Whether group
composition should be monitored instantaneously (truth), or sampled in
discrete time.}

\item{samples_per_day}{Only used in discrete time approximations in simulate_graph.}
}
\value{
sim_igraph, an igraph graph object.
}
\description{
Useful for those interested primarily in generating modular network graphs.
Equivalent to using simulate_schedule with the independent simulator and then
graph_from_schedule, but also includes a discrete-time approximation that is
faster for large networks.
}
\examples{
g <- simulate_graph(n_animals = 25,
                    n_groups = 4,
                    time_to_leave = 5,
                    time_to_return = 2,
                    travel_time = c(0.001,0.2),
                    sampling_duration = 7,
                    sampler = "discrete",
                    samples_per_day = 1
                    )
}
\seealso{
\code{\link{plot_simulated_graph}}
}
