% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_groups.R
\name{simulate_groups}
\alias{simulate_groups}
\title{Simulate groups}
\usage{
simulate_groups(
  animals_home,
  n_groups,
  time_to_leave,
  time_to_return,
  travel_time = c(0, 2),
  sampling_duration = 7,
  samples_per_day = 1
)
}
\description{
Simulate groups
}
