% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_independence.R
\name{simulate_independence}
\alias{simulate_independence}
\title{simulate group switching to get animals_transformed df}
\usage{
simulate_independence(
  n_animals,
  n_groups,
  time_to_leave,
  time_to_return,
  travel_time,
  sampling_duration
)
}
\arguments{
\item{n_groups}{the number of social groups and/or distinct patches in the population}

\item{time_to_leave}{the average number of days spent at home prior to leaving}

\item{time_to_return}{the average number of days spent away from home prior to returning}

\item{travel_time}{a list of two values in terms of days. These are the limits for draws from a uniform distribution governing how long it will take to get from one group to another.}

\item{sampling_duration}{the number of days to simulate.}
}
\description{
simulate group switching to get animals_transformed df
}
\keyword{internal}
