% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_non_independence.R
\name{simulate_non_independence}
\alias{simulate_non_independence}
\title{simulate non-independent group-switching}
\usage{
simulate_non_independence(
  n_groups = 4,
  n_animals = 16,
  time_to_leave = 5,
  time_to_return = 2,
  travel_time = c(0.01, 2),
  sampling_duration = 7
)
}
\arguments{
\item{n_groups}{The number of modules in the network.}

\item{time_to_leave}{The average amount of time spent within the home
group prior to leaving (days per sampling period).}

\item{time_to_return}{The average amount of time spent abroad before
returning to home group.}

\item{travel_time}{Vector of the range of values that travel to the next group can take,
drawn from a uniform distribution. Default == c(0,1)}

\item{sampling_duration}{Default == 365.}
}
\description{
simulate non-independent group-switching
}
\examples{
\donttest{
t2 <- simulate_non_independence(n_groups = 4,
time_to_leave = 5,
time_to_return = 2,
travel_time = c(0.01,2),
sampling_duration = 7,
samples_per_day = 1)
}
}
