% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stash.R
\name{stash}
\alias{stash}
\alias{unstash}
\alias{list_stashes}
\alias{remove_stash}
\title{Stash and Unstash Modules. List and Remove Stashes.}
\usage{
stash(comment = NA_character_)

unstash(id = length(.modulr_env$injector$stash))

list_stashes()

remove_stash(id, all)
}
\arguments{
\item{comment}{A string (character vector of length one).}

\item{id}{A scalar (integer vector of length one).}

\item{all}{A flag. Should all stashes be removed?}
}
\value{
A scalar (integer vector of length one) containing the stash ID.
}
\description{
Stash and unstash all modules and internal state. List and remove stashes.
}
\details{
\code{stash} allows to stash all modules and the internal state of modulr.
Each stash is pushed on a stack and can be commented for later reference.
\code{unstash} pops a stash from the stack and restores the state
accordingly. \code{list_stashes} lists the stack, showing ids, timestamps and
comments for each stash. \code{remove_stash} removes one or all stashes from
the stack.
}
\section{Warning}{

 It is considered a very bad practice to define, touch, undefine, load, make,
 reset, or perform any other operation from within a module definition that
 may alterate the internal state of modulr.
}

\examples{
reset()
remove_stash(all = TRUE)
define("foo", NULL, function() NULL)
stash("'foo' only")
list_stashes()
define("bar", NULL, function() NULL)
stash("'foo' and 'bar'")
list_stashes()
list_modules()
unstash(1L)
list_modules()

}
\seealso{
\code{\link{define}}, \code{\link{list_modules}}, and
  \code{\link{reset}}.
}
