% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_animals.R
\name{switch_animals}
\alias{switch_animals}
\title{Internal to simulate_non_independence}
\usage{
switch_animals(inlist, elem_leave)
}
\arguments{
\item{inlist}{a list like mbrs_list}

\item{elem_leave}{a list with same length  as mbrs_list of animals to send away
from their home group}
}
\description{
Sends animals leaving home to another group at current time step
}
\keyword{internal}
