% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_dist.R
\name{wgcna_dist}
\alias{wgcna_dist}
\title{Topological Overlap Matrix (TOM) Dissimilarity}
\usage{
wgcna_dist(object, params)
}
\arguments{
\item{object}{wide matrix or harmonized long object}

\item{params}{list of parameters for WGCNA routines}
}
\value{
numeric TOM dissimilarity matrix
}
\description{
Computes TOM dissimilarity matrix from trait matrix or harmonized object.
}
