% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_harmonize.R
\name{wgcna_harmonize}
\alias{wgcna_harmonize}
\title{Harmonize WGCNA module}
\usage{
wgcna_harmonize(
  mod = "WGCNAmodule",
  moddir = ".",
  modRdata = NULL,
  params = NULL,
  annot = NULL,
  harmonizeddir = ".",
  force = FALSE
)
}
\arguments{
\item{mod}{name for module object}

\item{moddir}{director}

\item{modRdata}{file name ending with `.Rdata`}

\item{params}{list of parameters}

\item{annot}{annotation file (ignored if `NULL`)}

\item{harmonizeddir}{name of directory to save `RDS` object in its `moddir`}

\item{force}{force creation if `TRUE`}
}
\value{
invisible file name for created object
}
\description{
This loads `geneTree`, `kMEs`, `merge`, `params` and `dynamicColors_2`
from an `Rdata` object and creates and `RDS` object for later reuse.
}
