% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_params.R
\name{wgcna_params}
\alias{wgcna_params}
\title{Parameters of WGCNA object}
\usage{
wgcna_params(params = NULL)
}
\arguments{
\item{params}{list of updated parameters (defaults if `NULL`)}
}
\value{
list with parameters
}
\description{
Parameters of WGCNA object
}
