% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_pivot.R
\name{wgcna_pivot}
\alias{wgcna_pivot}
\title{Pivot WGCNA object}
\usage{
wgcna_pivot(object)
}
\arguments{
\item{object}{of class `wgcnaModules`}
}
\value{
pivoted object in wide format for WGCNA routines
}
\description{
Pivot WGCNA object
}
