% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulr.R
\name{with_injector}
\alias{with_injector}
\title{With Injector.}
\usage{
with_injector(injector, code)
}
\arguments{
\item{injector}{An injector (R environment).}

\item{code}{Any object. Code to execute with the temporary injector.}
}
\value{
The result of the evaluation of the \code{code} argument.
}
\description{
Temporarily use a given injector.
}
\seealso{
\code{\link[withr]{withr}} for examples of 'with_' methods.
}
