% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with.R
\name{with_no_packages}
\alias{with_no_packages}
\alias{without_packages}
\title{With no packages.}
\usage{
with_no_packages(code, ignore = getOption("modulr.ignore_packages"))

without_packages(code, ignore = getOption("modulr.ignore_packages"))
}
\arguments{
\item{code}{Any object. Code to execute in the temporary environment.}

\item{ignore}{A vector of character strings. Packages (in addition to base
packages) to ignore in the process.}
}
\value{
The result of the evaluation of the \code{code} argument.
}
\description{
Temporarily unload and detach packages.
}
\details{
Packages are temporarily unloaded and detached from the search path. When
restored, the ordering of the packages on the search path is preserved. Base
packages and packages passed to the \code{ignore} parameter are ignored in
the process.
}
\examples{
with_no_packages({
  message("Look mum, no packages!")
  print(sessionInfo())
})

}
\seealso{
\code{\link[withr]{withr}} for examples of 'with_' methods,
  \code{\link{getOption}}, \code{\link{library}}, \code{\link{options}}, and
  \code{\link{search}}.
}

