% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulr.R
\name{with_verbosity}
\alias{with_verbosity}
\title{With verbosity.}
\usage{
with_verbosity(level, code)
}
\arguments{
\item{level}{A scalar (integer vector of length one),
possibly \code{-Inf} or \code{+Inf}.}

\item{code}{Any object. Code to execute in the temporary environment.}
}
\value{
The result of the evaluation of the \code{code} argument.
}
\description{
Temporarily change the verbosity.
}
\details{
Verbosity is temporarily changed.
}
\examples{
define("foo", NULL, { "bar" })
with_verbosity(-Inf, {
  make("foo")
})

}
\seealso{
\code{\link[withr]{withr}} for examples of 'with_' methods,
  \code{\link{get_verbosity}}, and \code{\link{set_verbosity}}.
}
