% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assortr_plot.R
\name{assortr_plot}
\alias{assortr_plot}
\title{Plot a graph, color-coded by \code{rnetcarto} modules}
\usage{
assortr_plot(x, alg = "walktrap")
}
\arguments{
\item{x, }{a named square adjacency matrix or named igraph graph. Intended for
weighted, undirected graphs.}

\item{alg, }{the community detection algorithm to use, either from igraph:
"fast_greedy", "leading_eigen", "louvain", or "walktrap"
or from rnetcarto: "netcarto"
default is "walktrap"}
}
\value{
An igraph plot using the Fruchterman-Reingold layout and displaying
color-coded node membership as determined by \code{rnetcarto}.
}
\description{
Plot a graph, color-coded by \code{rnetcarto} modules
}
\examples{
adjmat <- matrix(sample(seq(0, 1, .01), 16), nrow = 4)
diag(adjmat) <- 0
adjmat[lower.tri(adjmat)] <- 0
rownames(adjmat) <- colnames(adjmat) <- paste0("Animal_", 1:4)
assortr_plot(adjmat)
}
