% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bolts.R
\name{bolts}
\alias{bolts}
\title{internal function to graph crossing}
\usage{
bolts(animal, infectious)
}
\arguments{
\item{animal}{unique individual}

\item{infectious}{a data.table with infectious individuals from previous time step}
}
\description{
internal function to graph crossing
}
\keyword{internal}
