% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bolts.R
\name{bolts}
\alias{bolts}
\title{Internal Function to graph_crossing()}
\usage{
bolts(animal, infectious)
}
\arguments{
\item{animal}{unique individual}

\item{infectious}{a data.table with infectious individuals from previous time step}
}
\description{
Identifies instances where an individual was co-located with an infectious animal
}
\keyword{internal}
