% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_holding.R
\name{clean_holding}
\alias{clean_holding}
\title{Clean up the holding value strings}
\usage{
clean_holding(cell)
}
\arguments{
\item{cell}{the holding value string from a single cell}
}
\description{
Clean up the holding value strings
}
\keyword{internal}
