% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_vector.R
\name{clean_vector}
\alias{clean_vector}
\title{Internal function to simulate_non_independence2() that puts vector string in numeric order}
\usage{
clean_vector(cell)
}
\arguments{
\item{cell}{a t2$vector cell}
}
\description{
Internal function to simulate_non_independence2() that puts vector string in numeric order
}
\keyword{internal}
