% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/define.R
\name{define}
\alias{define}
\title{Define a module.}
\usage{
define(name, dependencies, factory, compress = "gzip")
}
\arguments{
\item{name}{the module name, given as a character string.}

\item{dependencies}{the list of module dependencies, given as module names.}

\item{factory}{the factory function.}
}
\value{
a wrapper function around the module instanciation.
}
\description{
Define a module.
}
\examples{
# define "module_1"
define("module_1", list(), function() {
 message("Module 1"); "value 1"})

# define "module_2"
m2 <- define("module_2", list("module_1"), function(m1) {
 message("Module 2 with one dependency"); paste(m1, "value 2")})

# instanciate "module_2"
m2()
}

