% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_fxn.R
\name{dt_fxn}
\alias{dt_fxn}
\title{Extract data from list on individual basis and make keyed data.table}
\usage{
dt_fxn(animal)
}
\arguments{
\item{animal}{an element of animal_list from simulate_graph}
}
\description{
Extract data from list on individual basis and make keyed data.table
}
