% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_group.R
\name{extract_group}
\alias{extract_group}
\title{wrapper on stringr::str_extract with regex for group tag}
\usage{
extract_group(x)
}
\arguments{
\item{x}{a string}
}
\description{
wrapper on stringr::str_extract with regex for group tag
}
