% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rates.R
\name{get_rates}
\alias{get_rates}
\title{Calculate averaged lambda and xi rates for all individuals in a network}
\usage{
get_rates(sched, sim)
}
\arguments{
\item{sched}{a schedule object for a network from simulate_schedule()}

\item{sim}{the simulator used to generate the schedule}
}
\description{
Wrapper to apply single_id_rate to all individuals in a network and calculate
the mean number of days individuals took to leave their respective home states
and the mean number of days individuals took to return to their respective
home states.
}
\examples{
\donttest{


}
}
