% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_times.R
\name{get_times}
\alias{get_times}
\title{get times}
\usage{
get_times(schedule, id, simulator)
}
\arguments{
\item{schedule}{a dataframe schedule of the type output by simulate_schedule}

\item{id}{the character string name of an individual node}
}
\description{
get times
}
