% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_times.R
\name{get_times}
\alias{get_times}
\title{Calculates the proportion of time an individual spends at home vs away}
\usage{
get_times(schedule, id, simulator, option = NA)
}
\arguments{
\item{schedule}{a dataframe schedule of the type output by simulate_schedule}

\item{id}{the character string name of an individual node}

\item{option}{additional option for group-think, "co-located" or "attached"}
}
\description{
Calculates the proportion of time an individual spends at home vs away
}
