% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_crossing.R
\name{graph_crossing}
\alias{graph_crossing}
\title{Recursive function to calculate graph crossing time and do agent-based SEIR modeling}
\usage{
graph_crossing(animals_transformed, exposure_time, infectious_time, index_case)
}
\arguments{
\item{exposure_time}{E}

\item{infectious_time}{I}

\item{index_case}{the animal to start the transmission chain with}
}
\description{
Recursive function to calculate graph crossing time and do agent-based SEIR modeling
}
