% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_from_non_independence.R
\name{graph_from_non_independence}
\alias{graph_from_non_independence}
\title{graph non independence}
\usage{
graph_from_non_independence(t2)
}
\arguments{
\item{t2}{a dataframe from simulate_non_independence()
#' @examples
\donttest{
t2 <- simulate_non_independence(n_groups = 4,
time_to_leave = 5,
time_to_return = 2,
cohesion = 0.8,
travel_time = c(0,2),
sampling_duration = 7,
samples_per_day = 1)
g <- graph_from_non_independence(t2)
}}
}
\description{
graph non independence
}
