% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_from_schedule.R
\name{graph_from_schedule}
\alias{graph_from_schedule}
\title{Create undirected, weighted 'igraph' graph from a continuous-time description of animal movements.}
\usage{
graph_from_schedule(schedule)
}
\arguments{
\item{schedule}{object from simulate_schedule()}
}
\value{
an undirected, weighted 'igraph' graph where weights are the proportion
of time dyads spent together.
}
\description{
Create undirected, weighted 'igraph' graph from a continuous-time description of animal movements.
}
\examples{
\donttest{
g <- graph_from_schedule(schedule = obj)
}
}
