% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_from_schedule.R
\name{graph_from_schedule}
\alias{graph_from_schedule}
\title{Convert Continuous-Time Movement Descriptions to Network Graph}
\usage{
graph_from_schedule(schedule)
}
\arguments{
\item{schedule}{object from simulate_schedule()}
}
\value{
an undirected, weighted 'igraph' graph
}
\description{
Takes a list of continuous-time movement descriptions for each individual in
a network and returns an undirected, weighted 'igraph' graph where edge weights
are the proportion of time a dyad spent co-located at a patch.
}
\examples{
\donttest{
g <- graph_from_schedule(schedule = obj)
}
}
