% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_from_schedule.R
\name{graph_from_schedule}
\alias{graph_from_schedule}
\title{Simulate the true simple ratio index values for each dyad in a network
using output from animals_transformed().}
\usage{
graph_from_schedule(schedule)
}
\arguments{
\item{schedule}{output from the animals_transformed function}
}
\value{
sim_igraph, an igraph graph object.
}
\description{
Simulate the true simple ratio index values for each dyad in a network
using output from animals_transformed().
}
