% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{hit}
\alias{hit}
\title{Interactively Find, Make and Bind a Module.}
\usage{
hit(
  name,
  suffix = getOption("modulr.hit_suffix"),
  replace = TRUE,
  execute = FALSE
)
}
\arguments{
\item{name}{A string (character vector of lenght one).

  A module name can contain letters, figures and some special characters,
  namely \code{_}, \code{-}, and \code{/}. The latter is a namespace
  separator.

  Names containing \code{/mock/}, \code{/mocks/}, \code{/test/},
  \code{/tests/}, \code{/example/}, or \code{/examples/} have a special
  meaning related to code testing and examples.

  The name "modulr" corresponds to a special module and is therefore
  reserved.}

\item{suffix}{A string (character vector of lenght one).
Suffix added to variable names.}

\item{replace}{A flag. Should existing bindings be replaced?}

\item{execute}{A flag. Should binding be automatically made?}
}
\description{
Interactively find, make, and bind a module.
}
\details{
See \code{\link{make}} and \code{\link{module_options}}.
}
\section{Warning}{

 It is considered a very bad practice to define, touch, undefine, load, make,
 reset, or perform any other operation from within a module definition that
 may alterate the internal state of modulr.
}

\examples{
"foobar" \%provides\% "Hello World"
"foo" \%provides\% "Hello"
\dontrun{hit("foo")}
\dontrun{hit(foo)}
}
\seealso{
\code{\link{make}}.
}
