% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate_group.R
\name{initiate_group}
\alias{initiate_group}
\title{Internal to simulate_non_independence and simulate_non_independence2}
\usage{
initiate_group(x, n)
}
\arguments{
\item{x}{the group number}

\item{n}{the number of individuals in the group to generate}
}
\description{
Creates string vector for all individuals in a group
}
\keyword{internal}
