% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_function.R
\name{interval_function}
\alias{interval_function}
\title{Internal to simulate_non_independence}
\usage{
interval_function(groups, intervals)
}
\arguments{
\item{groups}{a pairwise set of groups}

\item{intervals}{the intervals dataframe}
}
\description{
Used to make matching time intervals across all individuals in a network
}
\keyword{internal}
