% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{maps_config}
\alias{maps_config}
\title{Maps configuration.}
\format{\preformatted{List of 4
 $ unset  :function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 10 12 14 3 12 3 10 14
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x55724e0> 
 $ set    :function (..., drop = T)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 15 10 31 3 10 3 15 31
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x55724e0> 
 $ get    :function (key)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 40 10 42 3 10 3 40 42
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x55724e0> 
 $ get_all:function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 32 14 39 3 14 3 32 39
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x55724e0> 
}}
\usage{
maps_config
}
\description{
Maps configuration.
}
\keyword{datasets}

