% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{maps_config}
\alias{maps_config}
\title{Maps configuration.}
\format{\preformatted{List of 4
 $ unset  :function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 8 12 12 3 12 3 8 12
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x27f0268> 
 $ set    :function (..., drop = TRUE)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 14 10 51 3 10 3 14 51
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x27f0268> 
 $ get    :function (key)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 70 10 80 3 10 3 70 80
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x27f0268> 
 $ get_all:function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 53 14 68 3 14 3 53 68
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x27f0268> 
}}
\usage{
maps_config
}
\description{
Maps configuration.
}
\keyword{datasets}

