% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modulr.R
\docType{data}
\name{maps_config}
\alias{maps_config}
\title{Maps configuration.}
\format{\preformatted{List of 3
 $ set    :function (..., drop = T)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 35 10 51 3 10 3 35 51
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x206c248> 
 $ get    :function (key)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 56 10 58 3 10 3 56 58
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x206c248> 
 $ get_all:function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 52 14 55 3 14 3 52 55
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x206c248> 
}}
\usage{
maps_config
}
\description{
Maps configuration.
}
\keyword{datasets}

