% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_summary.R
\name{module_summary}
\alias{module_summary}
\title{Summarize Module Count Across Networks}
\arguments{
\item{network_list}{a list of 'igraph' graph objects with community memberships.}
}
\value{
a data.frame where each row corresponds the the nth element of
the network_list.
}
\description{
Calculates nodes-per-module and modules-per-network on a list of graphs with defined memberships
}
\examples{
\donttest{
g1 <- simulate_graph(n_animals = 12, n_groups = 3, time_to_leave = 4, time_to_return = 1,
                     travel_time = c(0.001, 0.002), sampling_duration = 30,
                     sampler = "discrete", samples_per_day = 1)
g2 <- simulate_graph(n_animals = 12, n_groups = 3, time_to_leave = 4, time_to_return = 1,
                     travel_time = c(0.001, 0.002), sampling_duration = 30,
                     sampler = "discrete", samples_per_day = 1)
g_list <- list(g1, g2)
module_summary(g_list)
}
}
