% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_summary.R
\name{module_summary}
\alias{module_summary}
\title{Calculate nodes per module and modules per network statistics on a list of
graphs with defined membership}
\arguments{
\item{network_list}{a list of 'igraph' graph objects with community memberships.}
}
\value{
a data.frame where each row corresponds the the nth element of
the network_list.
}
\description{
Calculate nodes per module and modules per network statistics on a list of
graphs with defined membership
}
\examples{
\donttest{
data(real_networks)
module_stats <- module_summary(real_networks)
module_stats \%>\%
  `names<-`(c("n", "mean", "sd", "netSize")) \%>\%
  ggplot2::ggplot() +
  ggplot2::geom_pointrange(ggplot2::aes(x = n, y = mean, ymin = mean - sd, ymax= mean + sd,
                      color = as.factor(netSize)),
                  position = ggplot2::position_jitter(width = .4), lty = "dotted") +
  ggplot2::theme_bw() +
  ggplot2::ggtitle("") +
  ggplot2::labs(x = "Modules per Network", y = "Mean of Nodes per Module")+
  ggplot2::theme(legend.title = ggplot2::element_text(size = 20),
        legend.text = ggplot2::element_text(size = 20),
        axis.text = ggplot2::element_text(size = 20),
        axis.title = ggplot2::element_text(size = 20)) +
  ggplot2::guides(color = ggplot2::guide_legend(title = "Network Size"))
  }
}
