% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulr-package.R
\docType{package}
\name{modulr-package}
\alias{modulr}
\alias{modulr-package}
\title{modulr: Simulate Individual Movement Behaviors Underlying Modular Fission-Fusion Networks}
\description{
A system for stochastic simulation of individual movements that generate social networks with controllable 'modularity' and sampling simulators for power analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gavincotterill/modulr}
  \item Report bugs at \url{https://github.com/gavincotterill/modulr/issues}
}

}
\author{
\strong{Maintainer}: Gavin Cotterill \email{gcotterill2@unl.edu} (\href{https://orcid.org/0000-0002-1408-778X}{ORCID})

Authors:
\itemize{
  \item Kezia Manlove \email{kezia.manlove@usu.edu} (\href{https://orcid.org/0000-0002-7200-5236}{ORCID})
}

}
\keyword{internal}
