% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulr-package.R
\docType{package}
\name{modulr-package}
\alias{modulr}
\alias{modulr-package}
\title{modulr: Simulate Modular Fission-Fusion Networks and Sample Design}
\description{
'The `modulr` package provides functions to simulate group switching behavior and create weighted, undirected social networks based on the proportion of time dyads spend together. `modulr` also simulates sampling processes and gives users to the tools to assess network statistic estimation under diverse scenarios'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gavincotterill/modulr}
  \item Report bugs at \url{https://github.com/gavincotterill/modulr/issues}
}

}
\author{
\strong{Maintainer}: Gavin Cotterill \email{gcotterill2@unl.edu} (\href{https://orcid.org/0000-0002-1408-778X}{ORCID})

}
\keyword{internal}
