% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modulr.R
\docType{package}
\name{modulr}
\alias{RESERVED_NAMES}
\alias{modulr}
\alias{modulr-package}
\title{modulr -- Module Pattern and DI in R}
\format{\preformatted{ chr "modulr"
}}
\usage{
RESERVED_NAMES
}
\description{
modulr is a Module Pattern and Dependency Injection implementation in R.
Module Pattern and DI allows you to encapsulate pieces of code into useful singleton units,
namely modules that register their capabilities, export values and rely on other modules as dependencies.
modulr is widely inspired from RequireJS and AngularJS for Javascript.
}
\author{
Alain Clément-Pavon <\email{alain.clement-pavon@unil.ch}>
}
\keyword{datasets}

