% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulr.R
\docType{package}
\name{modulr}
\alias{modulr}
\title{Modulr: A Dependency Injection Framework for R}
\description{
This package contains a Dependency Injection framework for R. It allows to
organize programs into discrete, modular, and loosely coupled units. By
nature, such modules are easy to develop, debug, test, reuse, share and
maintain in a wide range of common situations. This package may also allow
for higher compliance with best practices in software engineering.
}
\details{
To learn more about modulr, start with the vignettes and use
\code{browseVignettes(package = "modulr")}.

For a complete list of functions,
use \code{library(help = "modulr")}.
}
\author{
Alain Clément <\email{alain.clement@unil.ch}>
}
