% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcarto_modules.R
\name{netcarto_modules}
\alias{netcarto_modules}
\title{Return vector of group membership}
\usage{
netcarto_modules(x)
}
\arguments{
\item{x}{a named square adjacency matrix or named \code{igraph} graph. Intended for
weighted, undirected graphs.}
}
\value{
a named vector of module memberships matching the named adjacency
matrix or igraph graph object (igraph::V(graph)$name)
}
\description{
Return vector of group membership
}
\examples{
adjmat <- matrix(sample(seq(0, 1, .01), 16), nrow = 4)
diag(adjmat) <- 0
adjmat[lower.tri(adjmat)] <- 0
rownames(adjmat) <- colnames(adjmat) <- paste0("Animal_", 1:4)
netcarto_modules(adjmat)
}
