% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modulr.R
\docType{data}
\name{paths_config}
\alias{paths_config}
\title{Paths configuration.}
\format{\preformatted{List of 3
 $ set    :function (..., drop = T)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 31 10 47 3 10 3 31 47
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x1d3f5e8> 
 $ get    :function (key)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 52 10 54 3 10 3 52 54
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x1d3f5e8> 
 $ get_all:function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 48 14 51 3 14 3 48 51
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x1d3f5e8> 
}}
\usage{
paths_config
}
\description{
Paths configuration.
}
\keyword{datasets}

