% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modulr.R
\docType{data}
\name{paths_config}
\alias{paths_config}
\title{Paths configuration.}
\format{\preformatted{List of 3
 $ set    :function (..., drop = T)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 36 10 52 3 10 3 36 52
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x3330428> 
 $ get    :function (key)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 57 10 59 3 10 3 57 59
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x3330428> 
 $ get_all:function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 53 14 56 3 14 3 53 56
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x3330428> 
}}
\usage{
paths_config
}
\description{
Paths configuration.
}
\keyword{datasets}

