% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_dependencies}
\alias{graph_dependencies}
\alias{plot_dependencies}
\title{Plot Dependencies.}
\usage{
plot_dependencies(group, regexp, include_group = TRUE, reserved = TRUE, ...)
}
\arguments{
\item{group}{A character vector of module names (cf. \code{\link{define}}) to
include as a subset of the graph nodes.}

\item{regexp}{A regular expression. If not missing, the regular expression is
used to filter the names of the modules to be plotted.}

\item{include_group}{A flag. Should the group of modules be included to the
plot?}

\item{reserved}{A flag. Should special modules with a reserved name be
considered?}

\item{...}{For \code{make}, further arguments to be passed for evaluation to
the resulting function, if any. For \code{make_all_tests}, further
arguments to be passed to \code{\link{load_all_modules}}.}
}
\description{
Plot the directed acyclic graph (DAG) of modules and dependencies.
}
\examples{
reset()
define("foo", NULL, function() NULL)
define("bar", list(m = "modulr"), function(m) NULL)
define("foobar", list(f = "foo", b = "bar"), function(f, b) NULL)
define("foobuzz", list(f = "foo", b = "buzz"), function(f, b) NULL)
define("fizz", list(f1 = "foobar", f2 = "foobuz"), function(f1, f2) NULL)
wait <- function() invisible(readline(prompt="Press [enter] to continue"))
plot_dependencies("foobar", reserved = FALSE); wait()
plot_dependencies("foobar"); wait()
plot_dependencies(reserved = FALSE); wait()
plot_dependencies()

}
\seealso{
\code{\link{define}} and \code{\link{reset}}.
}

