% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_dependencies}
\alias{plot_dependencies}
\alias{graph_dependencies}
\title{Plot Dependencies.}
\usage{
plot_dependencies(
  group,
  regexp,
  reserved = TRUE,
  render_engine = sankey_engine,
  ...
)
}
\arguments{
\item{group}{A character vector of module names (cf. \code{\link{define}}) to
include as a subset of the graph nodes.}

\item{regexp}{A regular expression. If not missing, the regular expression is
used to filter the names of the modules to be plotted.}

\item{reserved}{A flag. Should special modules with a reserved name be
considered?}

\item{render_engine}{A function. Rendering engine used to plot the
dependencies.}

\item{...}{Further arguments to be passed to \code{render_engine}.}
}
\description{
Plot the directed acyclic graph (DAG) of modules and dependencies.
}
\examples{
reset()
define("foo", NULL, function() NULL)
define("bar", list(m = "modulr"), function(m) NULL)
define("foobar", list(f = "foo", b = "bar"), function(f, b) NULL)
define("foobuzz", list(f = "foo", b = "buzz"), function(f, b) NULL)
define("fizz", list(f1 = "foobar", f2 = "foobuz"), function(f1, f2) NULL)
wait <- function() invisible(readline(prompt="Press [enter] to continue"))
plot_dependencies("foobar", reserved = FALSE)
\dontrun{wait()}
plot_dependencies("foobar")
\dontrun{wait()}
plot_dependencies(reserved = FALSE)
\dontrun{wait()}
plot_dependencies()

}
\seealso{
\code{\link{define}} and \code{\link{reset}}.
}
