% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sampled_graph.R
\name{plot_sampled_graph}
\alias{plot_sampled_graph}
\title{Plot Graph of Sampled Network}
\usage{
plot_sampled_graph(
  g_obs,
  g,
  vertex.size = 40,
  mark.expand = 25,
  vertex.label = NA,
  vertex.label.cex = 1.5,
  title = "",
  seed = NULL
)
}
\arguments{
\item{g_obs}{an igraph object output from sample_graph()}

\item{g}{an igraph object output from simulate_graph() or graph_from_schedule()}

\item{vertex.size}{node size in plot}

\item{mark.expand}{padding of polygon denoting modules around nodes}

\item{vertex.label}{label for graph vertices}

\item{vertex.label.cex}{size for vertex labels}

\item{title}{plot title}

\item{seed}{optional integer value to set.seed() within function and preserve node layout when plotting sampled and simulated graphs}
}
\value{
a plot of the sampled graph
}
\description{
Takes a graph from sample_graph and plots it to match plot_simulated_graph.
If optional seed is set to match across the two functions, node position is preserved.
}
\examples{
\donttest{
library(igraph)
g <- modulr::simulate_graph(n_animals = 25,
                    n_groups = 4,
                    time_to_leave = 5,
                    time_to_return = 2,
                    samples_per_day = 1,
                    sampling_duration = 7)

g_obs <- sample_graph(
  graph = g,
  sample_nNodes = 13,
  prop_hi_res = 1,
  regime = "grab-two",
  alg = "fast_greedy")

plot_sampled_graph(g_obs, g)
}
}
