% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_simulated_graph.R
\name{plot_simulated_graph}
\alias{plot_simulated_graph}
\title{plot_simulated_graph()}
\usage{
plot_simulated_graph(
  g,
  vertex.size = 40,
  mark.expand = 25,
  vertex.label = NA,
  vertex.label.cex = 1.5,
  title = ""
)
}
\arguments{
\item{g, }{an igraph object output from simulate_graph()}

\item{vertex.size, }{node size in plot}

\item{mark.expand, }{padding of polygon denoting modules around nodes}

\item{vertex.label, }{label for graph vertices}

\item{vertex.label.cex, }{size for vertex labels}

\item{title, }{plot title}
}
\value{
a plot of the igraph object
}
\description{
plot_simulated_graph()
}
\examples{
set.seed(123)
g <- simulate_graph(n_animals = 25,
                    n_groups = 4,
                    time_to_leave = 5,
                    time_to_return = 2,
                    samples_per_day = 1,
                    sampling_duration = 7)
plot_simulated_graph(g)

}
\seealso{
\code{\link{simulate_graph}}
}
