% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_simulated_graph.R
\name{plot_simulated_graph}
\alias{plot_simulated_graph}
\title{plot_simulated_graph()}
\usage{
plot_simulated_graph(g, vertex.size = 40, mark.expand = 25)
}
\arguments{
\item{g, }{an igraph object output from simulate_graph()}
}
\value{

}
\description{
plot_simulated_graph()
}
\examples{
plot_simulated_graph(g)

}
\seealso{
\code{\link{simulate_graph, plot_sampled_graph}}
}
