% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_transmissions.R
\name{plot_transmissions}
\alias{plot_transmissions}
\title{plot_transmissions()}
\usage{
plot_transmissions(
  g,
  vertex.size = 40,
  mark.expand = 25,
  vertex.label = NA,
  edge.arrow.size = NA,
  vertex.label.cex = 1.5,
  title = ""
)
}
\arguments{
\item{g, }{an igraph object output from simulate_graph()}

\item{vertex.size, }{node size in plot}

\item{mark.expand, }{padding of polygon denoting modules around nodes}

\item{vertex.label, }{label for graph vertices}

\item{vertex.label.cex, }{size for vertex labels}

\item{title, }{plot title}
}
\value{
a plot of the igraph object
}
\description{
plot_transmissions()
}
