% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gears.R
\name{prepare_gear}
\alias{prepare_gear}
\title{Prepare Modulr Gear.}
\usage{
prepare_gear(name = .Last.name, url = NULL, load = TRUE)
}
\arguments{
\item{name}{A string (character vector of lenght one).

  A module name can contain letters, figures and some special characters,
  namely \code{_}, \code{-}, and \code{/}. The latter is a namespace
  separator.

  Names containing \code{/mock/}, \code{/mocks/}, \code{/test/},
  \code{/tests/}, \code{/example/}, or \code{/examples/} have a special
  meaning related to code testing and examples.

  The name "modulr" corresponds to a special module and is therefore
  reserved.}

\item{url}{A string (character vector of length one). The URL must use the
HTTP(S) protocol.}

\item{load}{A flag. Should an undefined module be implicitely loaded?}
}
\value{
A string (character vector of length one) containing a modulr gear R
  Markdown script.
}
\description{
Prepare a module (including mocks, tests and examples) for publication as a
modulr gear.
}
\details{
A modulr gear is an R Markdown script containing a module, its
\link{docstring} info, and its special modules definitions, like mocks,
tests, and examples, for instance. \code{prepare_gear} prepares such a Modulr
Gear. If an \code{url} is given, the script will be completed accordingly
(see example).
}
\examples{
reset()
define("foo", NULL, function() {
 #' # A `foo` module
 #'
 #' Description goes here.
 NULL
})
cat(prepare_gear("foo", url = "https://example.org/gears/foo"), sep = "\n")

}
\seealso{
\code{\link{define}}, \code{\link{release_gear_as_gist}}, \code{\link{info}},
  and \code{\link{reset}}.
}
