% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_rel.R
\name{q_rel}
\alias{q_rel}
\title{Calculate the Newman's Q}
\usage{
q_rel(graph_object)
}
\arguments{
\item{graph_object}{an \code{igraph} graph object with 'membership' attribute}
}
\description{
Wrapper around assortnet function that takes a weighted, undirected network graph and returns the assortativity coefficient.
}
\examples{
\donttest{
g <- simulate_graph(n_animals = 25,
                    n_groups = 4,
                    time_to_leave = 5,
                    time_to_return = 2,
                    travel_time = c(0.01, 0.2),
                    samples_per_day = 1,
                    sampling_duration = 7)
q_rel(g)
}
}
