% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{root_config}
\alias{root_config}
\title{Root configuration.}
\format{\preformatted{List of 4
 $ unset  :function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 8 12 16 3 12 3 8 16
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x3451280> 
 $ set    :function (..., drop = T)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 18 10 59 3 10 3 18 59
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x3451280> 
 $ get    :function (key)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 80 10 90 3 10 3 80 90
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x3451280> 
 $ get_all:function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 61 14 78 3 14 3 61 78
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x3451280> 
}}
\usage{
root_config
}
\description{
Root configuration.
}
\keyword{datasets}

