% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_graph.R
\name{sample_graph}
\alias{sample_graph}
\title{Simulate a sampling regime and calculate network modularity.}
\usage{
sample_graph(
  adjmat,
  graph,
  missingness,
  propGPS,
  gps_freq,
  vhf_freq,
  regime = "random"
)
}
\arguments{
\item{adjmat}{I'm not sure why this is here... should just build it from the
graph}

\item{graph}{The true network to sample}

\item{missingness}{The proportion of the nodes in the real network that will
be missed in sampling}

\item{propGPS}{The proportion of high resolution data to use}

\item{gps_freq}{The observation frequency (observations per sampling period)
of the high res data}

\item{vhf_freq}{Observations per sampling period of the low res data}

\item{regime}{Whether to sample modules randomly or impose 'evenness'}
}
\value{
g_obs, the graph of the observed network
}
\description{
Simulate a sampling regime and calculate network modularity.
}
\examples{
# nothing yet
}
