% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_graph.R
\name{sample_graph}
\alias{sample_graph}
\title{Simulate network sampling, monitoring and perform community detection based on the simple ratio index.}
\usage{
sample_graph(
  graph,
  sample_nNodes,
  sampling_duration,
  prop_hi_res = 1,
  hi_res = 1,
  lo_res = 1/7,
  regime = "grab-two",
  alg = "fast_greedy"
)
}
\arguments{
\item{graph}{The true network to sample.}

\item{sample_nNodes}{The number of individuals to sample.}

\item{sampling_duration}{The number of days to monitor sampled individuals.}

\item{prop_hi_res}{The proportion of sampled nodes to monitor with high resolution.}

\item{hi_res}{The observation frequency of the high resolution individuals, expressed as the number of observations per day.}

\item{lo_res}{The observation frequency of the low resolution individuals, expressed as the number of observations per day.}

\item{regime}{The sampling regime to use: options include "random", "even", "grab-two".}

\item{alg}{The community detection algorithm to use, one of: "netcarto", "fast_greedy", "label_prop", "leading_eigen", "louvain", "spinglass", "walktrap".}
}
\value{
An \code{igraph} graph of the sampled network where edge weights ('attr' = "weight") is the simple ratio index.
}
\description{
Simulate network sampling, monitoring and perform community detection based on the simple ratio index.
}
\examples{
\donttest{
library(igraph)
data("real_networks")
data("study_design")
g_obs <- sample_graph(
  graph = real_networks[[1]],
  sample_nNodes = ceiling(0.5 * length(igraph::V(real_networks[[1]]))),
  prop_hi_res = 1,
  regime = "grab-two",
  alg = "fast_greedy")

}
}
