% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_animal.R
\name{simulate_animal}
\alias{simulate_animal}
\title{The backbone of the network simulator, called from within simulate_graph}
\usage{
simulate_animal(
  n_groups,
  time_to_leave,
  time_to_return,
  travel_time = c(0.001, 0.002),
  sampling_duration,
  samples_per_day = 1
)
}
\arguments{
\item{n_groups}{The number of modules in the network.}

\item{time_to_leave}{The average amount of time spent within the home
group prior to leaving (days per sampling period).}

\item{time_to_return}{The average amount of time spent abroad before
returning to home group.}

\item{travel_time}{Vector of the range of values that travel to the next group can take,
drawn from a uniform distribution. Default == c(0,1)}

\item{sampling_duration}{Default == 365.}

\item{samples_per_day}{Default == 1.}
}
\value{
A vector of sampling locations for a single animal.
}
\description{
The backbone of the network simulator, called from within simulate_graph
}
\keyword{internal}
