% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_animal.R
\name{simulate_animal}
\alias{simulate_animal}
\title{The backbone of the network simulator, called from within simulate_graph}
\usage{
simulate_animal(
  n_groups,
  time_to_leave,
  time_to_return,
  sampling_duration = 365,
  samples_per_day = 1
)
}
\arguments{
\item{n_groups}{The number of modules in the network.}

\item{time_to_leave}{The average amount of time spent within the home
group prior to leaving (days per sampling period).}

\item{time_to_return}{The average amount of time spent abroad before
returning to home group.}

\item{sampling_duration}{Default == 365.}

\item{samples_per_day}{Default == 1.}
}
\value{
A vector of sampling locations for a single animal.
}
\description{
The backbone of the network simulator, called from within simulate_graph
}
\examples{
simulate_animal(time_to_leave = 3,
time_to_return = 1,
n_groups = 4,
samples_per_day = 1,
sampling_duration = 7)
}
