% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_graph.R
\name{simulate_graph}
\alias{simulate_graph}
\title{Simulate the true simple ratio index values for each dyad in a network.}
\usage{
simulate_graph(
  n_animals,
  n_groups,
  time_to_leave,
  time_to_return,
  sampling_duration,
  sampler = "discrete",
  samples_per_day = 1
)
}
\arguments{
\item{n_animals}{The number of nodes to include in the network.}

\item{n_groups}{The number of modules in the network.}

\item{time_to_leave}{The average number of days spent in the home group.}

\item{time_to_return}{The average number of days spent away from the home
group.}

\item{sampling_duration}{The number of days in the sampling period.}

\item{sampler}{Takes one of c("discrete", "continuous"). Whether group
composition should be monitored instantaneously (truth), or sampled in
discrete time. Continuous monitoring is much slower.}

\item{samples_per_day}{Optional. If sampler == "discrete", the number of
samples taken per day in order to estimate the simple ratio index for each
dyad.}
}
\value{
an igraph graph object.
}
\description{
Simulate the true simple ratio index values for each dyad in a network.
}
\examples{
\donttest{
set.seed(123)
g <- simulate_graph(n_animals = 25,
                    n_groups = 4,
                    time_to_leave = 5,
                    time_to_return = 2,
                    sampling_duration = 7,
                    sampler = "discrete",
                    samples_per_day = 1
                    )
}
}
