% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_graph.R
\name{simulate_graph}
\alias{simulate_graph}
\title{Simulate the true simple ratio index values for each dyad in a network.}
\usage{
simulate_graph(
  n_animals,
  n_groups,
  time_to_leave,
  time_to_return,
  samples_per_day,
  sampling_duration,
  time_cut
)
}
\arguments{
\item{n_animals}{The number of nodes to include in the network.}

\item{n_groups}{The number of modules in the network.}

\item{time_to_leave}{The average number of days spent in the home group.}

\item{time_to_return}{The average number of days spent away from the home
group.}

\item{samples_per_day}{The number of samples per day.}

\item{sampling_duration}{The number of days in the sampling period.}

\item{time_cut}{An optional parameter indicating whether or not to simulate
values at a time period prior to the full sampling duration.... does this
actaully make sense here? Or should it just exist in sim_animal??}
}
\value{
sim_igraph, an igraph graph object.
}
\description{
Simulate the true simple ratio index values for each dyad in a network.
}
\examples{
set.seed(123)

g <- simulate_graph(n_animals = 25,
                    n_groups = 4,
                    time_to_leave = 5,
                    time_to_return = 2,
                    samples_per_day = 1,
                    sampling_duration = 7,
                    time_cut = 7)
}
\seealso{
\code{\link{plot_simulated_graph}}
}
