% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_graph.R
\name{simulate_graph}
\alias{simulate_graph}
\title{Simulate weighted, undirected network graph with independent switching assumptions and simple ratio index.}
\usage{
simulate_graph(
  n_animals,
  n_groups,
  time_to_leave,
  time_to_return,
  travel_time,
  sampling_duration,
  sampler = "discrete",
  samples_per_day = 1
)
}
\arguments{
\item{n_animals}{The number of nodes to include in the network.}

\item{n_groups}{The number of modules in the network.}

\item{time_to_leave}{The average amount of time spent within the home
group prior to leaving (days per sampling period).}

\item{time_to_return}{The average amount of time spent abroad before
returning to home group.}

\item{travel_time}{Vector of the range of values that travel to the next group can take,
drawn from a uniform distribution. Default == c(0,1)}

\item{sampling_duration}{Default == 365.}

\item{sampler}{Takes one of c("discrete", "continuous"). Whether group
composition should be monitored instantaneously (truth), or sampled in
discrete time.}

\item{samples_per_day}{Default == 1.}
}
\value{
sim_igraph, an igraph graph object.
}
\description{
Simulate weighted, undirected network graph with independent switching assumptions and simple ratio index.
}
\examples{
g <- simulate_graph(n_animals = 25,
                    n_groups = 4,
                    time_to_leave = 5,
                    time_to_return = 2,
                    travel_time = c(0.001,0.2),
                    sampling_duration = 7,
                    sampler = "discrete",
                    samples_per_day = 1
                    )
}
\seealso{
\code{\link{plot_simulated_graph}}
}
