% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_groups.R
\name{simulate_groups}
\alias{simulate_groups}
\title{Simulate groups}
\usage{
simulate_groups(
  animals_home,
  n_groups,
  time_to_leave,
  time_to_return,
  travel_time = c(0.001, 2),
  sampling_duration = 7
)
}
\arguments{
\item{animals_home}{an integer value representing an individuals home or preferred patch}

\item{n_groups}{the number of social groups and/or distinct patches in the population}

\item{time_to_leave}{the average number of days spent at home prior to leaving}

\item{time_to_return}{the average number of days spent away from home prior to returning}

\item{travel_time}{a list of two values in terms of days. These are the limits for draws from a uniform distribution governing how long it will take to get from one group to another.}

\item{sampling_duration}{the number of days to simulate.}
}
\description{
Simulate groups
}
\keyword{internal}
