% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_groups2.R
\name{simulate_groups2}
\alias{simulate_groups2}
\title{Simulate groups experimental for use in sim_non_ind2}
\usage{
simulate_groups2(
  animals_home,
  n_groups,
  n_splits,
  time_to_leave,
  time_to_return,
  travel_time = c(0.001, 0.001),
  sampling_duration
)
}
\arguments{
\item{n_groups}{The number of modules in the network.}

\item{n_splits}{the number of splinter groups that leave home, one or more}

\item{time_to_leave}{The average amount of time spent within the home
group prior to leaving (days per sampling period).}

\item{time_to_return}{The average amount of time spent abroad before
returning to home group.}

\item{travel_time}{Vector of the range of values that travel to the next group can take,
drawn from a uniform distribution. Default == c(0,1)}

\item{sampling_duration}{Default == 365.}
}
\description{
Simulate groups experimental for use in sim_non_ind2
}
\keyword{internal}
