% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_non_independence.R
\name{simulate_non_independence}
\alias{simulate_non_independence}
\title{Internal to simulate_schedule}
\usage{
simulate_non_independence(
  n_groups = 4,
  n_animals = 16,
  time_to_leave = 5,
  time_to_return = 2,
  travel_time = c(0.01, 0.02),
  sampling_duration = 7
)
}
\arguments{
\item{n_groups}{the number of social groups and/or distinct patches in the population}

\item{time_to_leave}{the average number of days spent at home prior to leaving}

\item{time_to_return}{the average number of days spent away from home prior to returning}

\item{travel_time}{a list of two values in terms of days. These are the limits for draws from a uniform distribution governing how long it will take to get from one group to another.}

\item{sampling_duration}{the number of days to simulate.}
}
\description{
sim = "group-think"
}
\keyword{internal}
