% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_id_rate.R
\name{single_id_rate}
\alias{single_id_rate}
\title{get the lambda and xi rates for a single individual}
\usage{
single_id_rate(sub_sched, sub_id, sim)
}
\arguments{
\item{sub_sched}{the individual's continuous time description}

\item{sub_id}{the individual's id string}

\item{sim}{the simulator used to generate the schedule}
}
\description{
get the lambda and xi rates for a single individual
}
\keyword{internal}
