% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_topology.R
\name{wgcna_topology}
\alias{wgcna_topology}
\alias{ggplot_wgcna_topology}
\title{WGCNA topology}
\usage{
wgcna_topology(
  object,
  powers = c(c(1:10), seq(from = 12, to = 20, by = 2)),
  verbose = 0,
  ...
)

ggplot_wgcna_topology(object, cutoff = 0.9, ...)
}
\arguments{
\item{object}{object of class}

\item{powers}{vector of beta values}

\item{verbose}{level of verbose messages}

\item{...}{}

\item{cutoff}{explained variation cutoff}
}
\value{
data frame of class `wgcna_topology`

ggplot object
}
\description{
WGCNA topology

GGplot of WGCNA Topology
}
