% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_wgcna.R, R/wgcna_modules.R
\name{ggplot_wgcnaModules}
\alias{ggplot_wgcnaModules}
\alias{autoplot.wgcnaModules}
\alias{wgcnaModules}
\alias{listof_wgcnaModules}
\alias{summary.listof_wgcnaModules}
\alias{summary_wgcnaModules}
\alias{summary.wgcnaModules}
\alias{plot_wgcnaModules}
\alias{plot.wgcnaModules}
\title{GGplot of WGCNA Modules}
\usage{
ggplot_wgcnaModules(object, main = "Gene dendrogram and module colors", ...)

\method{autoplot}{wgcnaModules}(object, ...)

wgcnaModules(object, params = NULL)

listof_wgcnaModules(traitData, traitSignal)

\method{summary}{listof_wgcnaModules}(object, ...)

summary_wgcnaModules(object, ...)

\method{summary}{wgcnaModules}(object, ...)

plot_wgcnaModules(x, main = "Gene dendrogram and module colors", ...)

\method{plot}{wgcnaModules}(x, ...)
}
\arguments{
\item{object}{object of class `listof_wgcnaModules`}

\item{main}{title for plot}

\item{...}{additional parameters}

\item{params}{list of parameters for WGCNA routines}

\item{traitData}{data frame from `foundr::traitData()`}

\item{traitSignal}{data frame from `foundr::partition()`}

\item{object, x}{object of class `wgcnaModules`}
}
\value{
ggplot2 object

ggplot2 object

object of class wgcnaModules

object of class `listof_wgcnaModules`

data frame
}
\description{
Data in `object` are assumed to be in long format with columns
a subset of "dataset", "strain", "sex", "animal", "condition"
}
