% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_modules.R
\name{wgcnaModules}
\alias{wgcnaModules}
\alias{summary_wgcnaModules}
\alias{summary.wgcnaModules}
\alias{plot_wgcnaModules}
\alias{plot.wgcnaModules}
\title{Create WGCNA Modules}
\usage{
wgcnaModules(object, params = NULL)

summary_wgcnaModules(object, ...)

\method{summary}{wgcnaModules}(object, ...)

plot_wgcnaModules(x, main = "Gene dendrogram and module colors", ...)

\method{plot}{wgcnaModules}(x, ...)
}
\arguments{
\item{object}{harmonized data frame from routine `foundr`}

\item{params}{list of parameters for WGCNA routines}

\item{...}{additional parameters}

\item{main}{title for plot}

\item{object, x}{object of class `wgcnaModules`}
}
\value{
object of class wgcnaModules
}
\description{
Data in `object` are assumed to be in long format with columns
a subset of "dataset", "strain", "sex", "animal", "condition"
}
