% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_modules.R
\name{wgcnaModules}
\alias{wgcnaModules}
\alias{listof_wgcnaModules}
\alias{plot_wgcnaModules}
\alias{plot.wgcnaModules}
\title{Create WGCNA Modules}
\usage{
wgcnaModules(object, params = NULL)

listof_wgcnaModules(traitData, traitSignal, params = NULL)

plot_wgcnaModules(x, main = "Gene dendrogram and module colors", ...)

\method{plot}{wgcnaModules}(x, ...)
}
\arguments{
\item{object}{harmonized data frame from routine `foundr`}

\item{params}{list of parameters for WGCNA routines}

\item{traitData}{data frame from `foundr::traitData()`}

\item{traitSignal}{data frame from `foundr::partition()`}
}
\value{
object of class wgcnaModules

object of class `listof_wgcnaModules`
}
\description{
Data in `object` are assumed to be in long format with columns
a subset of "dataset", "strain", "sex", "animal", "condition"
}
